# Conflicts:
#	lib/app/bloc/Stream/ProductBloc.dart
#	lib/app/model/api/APIRepository.dart
#	lib/app/model/api/_APIProvider.dart
#	lib/app/model/core/AppRoute.dart
#	lib/app/model/pojo/response/ProducItemRespone.dart
#	lib/app/model/pojo/response/ProductRespone.dart
#	lib/app/model/pojo/response/WishlistsRespone.dart
#	lib/app/ui/me/myproduct/productDetailShop/ProductDetailShop.dart
#	lib/app/ui/me/myshop/myorderhistory/review/ReviewView.dart
#	lib/app/ui/order/OrderView.dart
#	lib/app/ui/productdetail/ProductDetailView.dart
#	lib/app/ui/productdetail/widget/ProductInto.dart
#	lib/app/ui/productdetail/widget/Reviewscore.dart
#	lib/app/ui/wishlist/WishlistsView.dart
#	lib/utility/widgets/ProductItemCard.dart
#	lib/utility/widgets/ProductVertical.dart

+++ lib/app/model/pojo/response/ProductMyShopRespone.dart
@@ -46,7 +46,7 @@ class ProductMyShopRespone {
   List<ImageProductShop> image;
   int discountPercent;
   int rating;
-  double reviewCount;
+  int reviewCount;
 
   ProductMyShopRespone(
       {this.id,
